Add outputtype module
Moves OutputType into its own module
