+++ src/wallet/init.cpp
@@ -7,6 +7,7 @@
 #include <init.h>
 #include <net.h>
 #include <scheduler.h>
+#include <outputtype.h>
 #include <util.h>
 #include <utilmoneystr.h>
 #include <validation.h>
